+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `/user` slash command
 - Added field to retaliation embed for whether the original attack is a retaliation
 
+### Changed
+- Re-enabled Discord-based authentication
+
 ### Fixed
 - Fixed retaliations not being marked as completed and being spammed